backend/requirements.txt
sqlalchemy
psycopg2-binary
alembic
fastapi
uvicorn
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
fpdf2
loguru
slowapi
python-dotenv
firebase-admin
